requirements.txt
Flask
flask-cors
dnspython
requests